phpcs fixes

+++ src/bp-media/bp-media-settings.php
@@ -1906,7 +1906,6 @@ function bb_media_settings_callback_symlink_support() {
     </label>
 
     <?php
-
 	$has_error = false;
 	if ( function_exists( 'ini_get' ) && ini_get( 'disable_functions' ) ) {
 
@@ -1923,7 +1922,6 @@ function bb_media_settings_callback_symlink_support() {
 			</div>
 			<?php
 		}
-
 	}
 
 	if ( empty( $has_error ) && bb_enable_symlinks() && empty( bp_get_option( 'bb_media_symlink_type' ) ) ) {
@@ -1975,88 +1973,81 @@ function bb_media_settings_callback_symlink_direct_access() {
 	$media_attachment_id    = 0;
 	$document_attachment_id = 0;
 
-    // Add upload filters.
-    add_filter( 'upload_dir', 'bp_video_upload_dir_script' );
-
+	// Add upload filters.
+	add_filter( 'upload_dir', 'bp_video_upload_dir_script' );
 	$file        = buddypress()->plugin_dir . 'bp-core/images/suspended-mystery-man.jpg';
 	$filename    = basename( $file );
 	$upload_file = wp_upload_bits( $filename, null, file_get_contents( $file ) );
 
-    if ( ! $upload_file['error'] ) {
-        $wp_filetype         = wp_check_filetype( $filename, null );
-        $attachment          = array(
-            'post_mime_type' => $wp_filetype['type'],
-            'post_title'     => preg_replace( '/\.[^.]+$/', '', $filename ),
-            'post_content'   => '',
-            'post_status'    => 'inherit',
-        );
-        $video_attachment_id = wp_insert_attachment( $attachment, $upload_file['file'] );
-        if ( ! is_wp_error( $video_attachment_id ) ) {
-            require_once ABSPATH . 'wp-admin/includes/image.php';
-            $attachment_data = wp_generate_attachment_metadata( $video_attachment_id, $upload_file['file'] );
-            wp_update_attachment_metadata( $video_attachment_id, $attachment_data );
-            $get_sample_ids['bb_videos'] = $video_attachment_id;
-        }
-    }
+	if ( ! $upload_file['error'] ) {
+		$wp_filetype         = wp_check_filetype( $filename, null );
+		$attachment          = array(
+			'post_mime_type' => $wp_filetype['type'],
+			'post_title'     => preg_replace( '/\.[^.]+$/', '', $filename ),
+			'post_content'   => '',
+			'post_status'    => 'inherit',
+		);
+		$video_attachment_id = wp_insert_attachment( $attachment, $upload_file['file'] );
+		if ( ! is_wp_error( $video_attachment_id ) ) {
+			require_once ABSPATH . 'wp-admin/includes/image.php';
+			$attachment_data = wp_generate_attachment_metadata( $video_attachment_id, $upload_file['file'] );
+			wp_update_attachment_metadata( $video_attachment_id, $attachment_data );
+			$get_sample_ids['bb_videos'] = $video_attachment_id;
+		}
+	}
+	// Remove upload filters.
+	remove_filter( 'upload_dir', 'bp_video_upload_dir_script' );
 
-    // Remove upload filters.
-    remove_filter( 'upload_dir', 'bp_video_upload_dir_script' );
+	add_filter( 'upload_dir', 'bp_media_upload_dir_script' );
+	$file        = buddypress()->plugin_dir . 'bp-core/images/suspended-mystery-man.jpg';
+	$filename    = basename( $file );
+	$upload_file = wp_upload_bits( $filename, null, file_get_contents( $file ) );
 
-    add_filter( 'upload_dir', 'bp_media_upload_dir_script' );
+	if ( ! $upload_file['error'] ) {
+		$wp_filetype         = wp_check_filetype( $filename, null );
+		$attachment          = array(
+			'post_mime_type' => $wp_filetype['type'],
+			'post_title'     => preg_replace( '/\.[^.]+$/', '', $filename ),
+			'post_content'   => '',
+			'post_status'    => 'inherit',
+		);
+		$media_attachment_id = wp_insert_attachment( $attachment, $upload_file['file'] );
+		if ( ! is_wp_error( $media_attachment_id ) ) {
+			require_once ABSPATH . 'wp-admin/includes/image.php';
+			$attachment_data = wp_generate_attachment_metadata( $media_attachment_id, $upload_file['file'] );
+			wp_update_attachment_metadata( $media_attachment_id, $attachment_data );
+			$get_sample_ids['bb_medias'] = $media_attachment_id;
+		}
+	}
+	remove_filter( 'upload_dir', 'bp_media_upload_dir_script' );
 
-    $file        = buddypress()->plugin_dir . 'bp-core/images/suspended-mystery-man.jpg';
-    $filename    = basename( $file );
-    $upload_file = wp_upload_bits( $filename, null, file_get_contents( $file ) );
+	add_filter( 'upload_dir', 'bp_document_upload_dir_script' );
+	$file        = buddypress()->plugin_dir . 'bp-core/images/suspended-mystery-man.jpg';
+	$filename    = basename( $file );
+	$upload_file = wp_upload_bits( $filename, null, file_get_contents( $file ) );
 
-    if ( ! $upload_file['error'] ) {
-        $wp_filetype         = wp_check_filetype( $filename, null );
-        $attachment          = array(
-            'post_mime_type' => $wp_filetype['type'],
-            'post_title'     => preg_replace( '/\.[^.]+$/', '', $filename ),
-            'post_content'   => '',
-            'post_status'    => 'inherit',
-        );
-        $media_attachment_id = wp_insert_attachment( $attachment, $upload_file['file'] );
-        if ( ! is_wp_error( $media_attachment_id ) ) {
-            require_once ABSPATH . 'wp-admin/includes/image.php';
-            $attachment_data = wp_generate_attachment_metadata( $media_attachment_id, $upload_file['file'] );
-            wp_update_attachment_metadata( $media_attachment_id, $attachment_data );
-            $get_sample_ids['bb_medias'] = $media_attachment_id;
-        }
-    }
-
-    remove_filter( 'upload_dir', 'bp_media_upload_dir_script' );
-
-    add_filter( 'upload_dir', 'bp_document_upload_dir_script' );
-
-    $file                   = buddypress()->plugin_dir . 'bp-core/images/suspended-mystery-man.jpg';
-    $filename               = basename( $file );
-    $upload_file            = wp_upload_bits( $filename, null, file_get_contents( $file ) );
-
-    if ( ! $upload_file['error'] ) {
-        $wp_filetype            = wp_check_filetype( $filename, null );
-        $attachment             = array(
-            'post_mime_type' => $wp_filetype['type'],
-            'post_title'     => preg_replace( '/\.[^.]+$/', '', $filename ),
-            'post_content'   => '',
-            'post_status'    => 'inherit',
-        );
-        $document_attachment_id = wp_insert_attachment( $attachment, $upload_file['file'] );
-        if ( ! is_wp_error( $document_attachment_id ) ) {
-            require_once ABSPATH . 'wp-admin/includes/image.php';
-            $attachment_data = wp_generate_attachment_metadata( $document_attachment_id, $upload_file['file'] );
-            wp_update_attachment_metadata( $document_attachment_id, $attachment_data );
-            $get_sample_ids['bb_documents'] = $document_attachment_id;
-        }
-    }
-
-    remove_filter( 'upload_dir', 'bp_document_upload_dir_script' );
-
+	if ( ! $upload_file['error'] ) {
+		$wp_filetype            = wp_check_filetype( $filename, null );
+		$attachment             = array(
+			'post_mime_type' => $wp_filetype['type'],
+			'post_title'     => preg_replace( '/\.[^.]+$/', '', $filename ),
+			'post_content'   => '',
+			'post_status'    => 'inherit',
+		);
+		$document_attachment_id = wp_insert_attachment( $attachment, $upload_file['file'] );
+		if ( ! is_wp_error( $document_attachment_id ) ) {
+			require_once ABSPATH . 'wp-admin/includes/image.php';
+			$attachment_data = wp_generate_attachment_metadata( $document_attachment_id, $upload_file['file'] );
+			wp_update_attachment_metadata( $document_attachment_id, $attachment_data );
+			$get_sample_ids['bb_documents'] = $document_attachment_id;
+		}
+	}
+	remove_filter( 'upload_dir', 'bp_document_upload_dir_script' );
 
 	$directory = array();
 	foreach ( $get_sample_ids as $id => $v ) {
 		$fetch = wp_remote_get( wp_get_attachment_image_url( $v ) );
-		if ( ! is_wp_error( $fetch ) && isset( $fetch['response']['code'] ) && 200 == $fetch['response']['code'] ) {
+		if ( ! is_wp_error( $fetch ) && isset( $fetch['response']['code'] ) && 200 === $fetch['response']['code'] ) {
 			$directory[] = $id;
 		}
 	}
@@ -2083,8 +2074,15 @@ function bb_media_settings_callback_symlink_direct_access() {
 		esc_html__( ' tutorial to configure your server.', 'buddyboss' )
 	);
 
-	wp_delete_attachment( $document_attachment_id, true );
-	wp_delete_attachment( $media_attachment_id, true );
-	wp_delete_attachment( $video_attachment_id, true );
+	if ( 0 !== $document_attachment_id && ! is_wp_error( $document_attachment_id ) ) {
+		wp_delete_attachment( $document_attachment_id, true );
+	}
 
+	if ( 0 !== $media_attachment_id && ! is_wp_error( $media_attachment_id ) ) {
+		wp_delete_attachment( $media_attachment_id, true );
+	}
+
+	if ( 0 !== $video_attachment_id && ! is_wp_error( $video_attachment_id ) ) {
+		wp_delete_attachment( $video_attachment_id, true );
+	}
 }
